[remove fluid] remove api Linear

+++ python/paddle/fluid/tests/unittests/test_gradient_clip.py
@@ -600,7 +600,7 @@ class TestDygraphGradientClipFP64(unittest.TestCase):
         with fluid.dygraph.guard():
             inputs = fluid.layers.uniform_random(
                 [16, 5], min=-10, max=10
-            ).astype('float64')
+            ).astype('float32')
             linear = paddle.nn.Linear(5, 5)
             out = linear(fluid.dygraph.to_variable(inputs))
             loss = fluid.layers.reduce_mean(out)
